GateServer：登陆Gate验证逻辑

+++ Code_Java/core/src/com/mokylin/bleach/core/isc/ServerIdDef.java
@@ -8,5 +8,5 @@ package com.mokylin.bleach.core.isc;
  **/
 public class ServerIdDef {
     /**规定：所有ServerID都必须为正整数*/
-    public static long InvalidServerID = 0;
+    public static int InvalidServerID = 0;
 }
